# docs/swagger.yaml
basePath: /
definitions:
...
  model.Role:
    properties:
      description:
        type: string
      flag:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
host: localhost:9090
info:
  contact:
    email: grezkl@protonmail.com
    name: grezkl
    url: http://github.com/grezkl
  description: e-gov server
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: e-gov server
  version: "1.0"
paths:
...
  /delRoleBatch/:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 角色编号数组
        in: path
        items:
          type: integer
        name: id
        required: true
        type: array
      responses: {}
      summary: 删除角色（批量）
      tags:
      - Role
...
  /deleteRoleById/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 角色编号
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      summary: 删除角色（单个）
      tags:
      - Role
...
  /getRoleMenu/{roleId}:
    get:
      consumes:
      - application/json
      parameters:
      - description: 角色编号
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      summary: 获取角色菜单
      tags:
      - Role
...
  /rolePage:
    get:
      consumes:
      - application/json
      parameters:
      - description: 页数
        in: query
        name: pageNum
        required: true
        type: integer
      - description: 页大小
        in: query
        name: pageSize
        required: true
        type: integer
      - description: 角色名称
        in: query
        name: name
        type: string
      responses: {}
      summary: 角色管理分页
      tags:
      - Role
...
  /saveRole:
    post:
      consumes:
      - application/json
      parameters:
      - description: 角色内容
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/model.Role'
      responses: {}
      summary: 保存角色内容
      tags:
      - Role
...
swagger: "2.0"